+++ ci/install_osx.sh
@@ -1,26 +1,20 @@
-brew tap homebrew/science
+brew tap dartsim/dart # for ipopt
 
 brew update > /dev/null
 
-PACKAGES='
-git
-cmake
-assimp
-dartsim/dart/fcl
-bullet --with-double-precision
-ode --with-libccd --with-double-precision
-flann
-boost
-eigen
-tinyxml
-tinyxml2
-homebrew/science/libccd
-nlopt
-ipopt
-ros/deps/urdfdom
-ros/deps/urdfdom_headers
-ros/deps/console_bridge
-open-scene-graph
-'
-
-brew install $PACKAGES | grep -v '%$'
+brew install git
+brew install cmake
+brew install assimp
+brew install fcl
+brew install bullet --with-double-precision
+brew install ode --with-libccd --with-double-precision
+brew install flann
+brew install boost
+brew install eigen
+brew install tinyxml
+brew install tinyxml2
+brew install libccd
+brew install nlopt
+brew install ipopt
+brew install urdfdom
+brew install open-scene-graph