+++ dart/collision/CollisionDetector.cpp
@@ -82,6 +82,13 @@ CollisionDetector::CollisionObjectManager::CollisionObjectManager(
 }
 
 //==============================================================================
+CollisionDetector*
+CollisionDetector::CollisionObjectManager::getCollisionDetector()
+{
+  return mCollisionDetector;
+}
+
+//==============================================================================
 CollisionDetector::
 ManagerForUnsharableCollisionObjects::ManagerForUnsharableCollisionObjects(
     CollisionDetector* cd)
@@ -117,7 +124,7 @@ void
 CollisionDetector::ManagerForUnsharableCollisionObjects
 ::CollisionObjectDeleter::operator()(CollisionObject* object) const
 {
-  mCollisionObjectManager->mCollisionDetector->notifyCollisionObjectDestroying(
+  mCollisionObjectManager->getCollisionDetector()->notifyCollisionObjectDestroying(
         object);
 
   delete object;
@@ -180,8 +187,8 @@ void
 CollisionDetector::ManagerForSharableCollisionObjects
 ::CollisionObjectDeleter::operator()(CollisionObject* object) const
 {
-  mCollisionObjectManager->mCollisionDetector->notifyCollisionObjectDestroying(
-        object);
+  mCollisionObjectManager->getCollisionDetector()
+      ->notifyCollisionObjectDestroying(object);
   mCollisionObjectManager->mCollisionObjectMap.erase(object->getShapeFrame());
 
   delete object;