+++ dart/dynamics/SoftMeshShape.cpp
@@ -88,9 +88,18 @@ Eigen::Matrix3d SoftMeshShape::computeInertia(double /*mass*/) const
   return Eigen::Matrix3d::Zero();
 }
 
-void SoftMeshShape::updateVolume()
+//==============================================================================
+void SoftMeshShape::updateBoundingBox() const
 {
   // TODO(JS): Not implemented.
+  mIsBoundingBoxDirty = false;
+}
+
+//==============================================================================
+void SoftMeshShape::updateVolume() const
+{
+  // TODO(JS): Not implemented.
+  mIsVolumeDirty = false;
 }
 
 void SoftMeshShape::_buildMesh()