+++ dart/dynamics/ZeroDofJoint.cpp
@@ -201,6 +201,19 @@ double ZeroDofJoint::getPositionLowerLimit(std::size_t /*_index*/) const
 }
 
 //==============================================================================
+void ZeroDofJoint::setPositionLowerLimits(
+    const Eigen::VectorXd& /*lowerLimits*/)
+{
+  // Do nothing
+}
+
+//==============================================================================
+Eigen::VectorXd ZeroDofJoint::getPositionLowerLimits() const
+{
+  return Eigen::VectorXd::Zero(0);
+}
+
+//==============================================================================
 void ZeroDofJoint::setPositionUpperLimit(std::size_t /*_index*/,
                                          double /*_position*/)
 {
@@ -211,6 +224,19 @@ void ZeroDofJoint::setPositionUpperLimit(std::size_t /*_index*/,
 double ZeroDofJoint::getPositionUpperLimit(std::size_t /*_index*/) const
 {
   return 0.0;
+}
+
+//==============================================================================
+void ZeroDofJoint::setPositionUpperLimits(
+    const Eigen::VectorXd& /*upperLimits*/)
+{
+  // Do nothing
+}
+
+//==============================================================================
+Eigen::VectorXd ZeroDofJoint::getPositionUpperLimits() const
+{
+  return Eigen::VectorXd::Zero(0);
 }
 
 //==============================================================================
@@ -293,6 +319,18 @@ double ZeroDofJoint::getVelocityLowerLimit(std::size_t /*_index*/) const
 }
 
 //==============================================================================
+void ZeroDofJoint::setVelocityLowerLimits(const Eigen::VectorXd& /*lowerLimits*/)
+{
+  // Do nothing
+}
+
+//==============================================================================
+Eigen::VectorXd ZeroDofJoint::getVelocityLowerLimits() const
+{
+  return Eigen::VectorXd::Zero(0);
+}
+
+//==============================================================================
 void ZeroDofJoint::setVelocityUpperLimit(std::size_t /*_index*/,
                                          double /*_velocity*/)
 {
@@ -303,6 +341,18 @@ void ZeroDofJoint::setVelocityUpperLimit(std::size_t /*_index*/,
 double ZeroDofJoint::getVelocityUpperLimit(std::size_t /*_index*/) const
 {
   return 0.0;
+}
+
+//==============================================================================
+void ZeroDofJoint::setVelocityUpperLimits(const Eigen::VectorXd& /*upperLimits*/)
+{
+  // Do nothing
+}
+
+//==============================================================================
+Eigen::VectorXd ZeroDofJoint::getVelocityUpperLimits() const
+{
+  return Eigen::VectorXd::Zero(0);
 }
 
 //==============================================================================
@@ -385,6 +435,18 @@ double ZeroDofJoint::getAccelerationLowerLimit(std::size_t /*_index*/) const
 }
 
 //==============================================================================
+void ZeroDofJoint::setAccelerationLowerLimits(const Eigen::VectorXd& /*lowerLimits*/)
+{
+  // Do nothing
+}
+
+//==============================================================================
+Eigen::VectorXd ZeroDofJoint::getAccelerationLowerLimits() const
+{
+  return Eigen::VectorXd::Zero(0);
+}
+
+//==============================================================================
 void ZeroDofJoint::setAccelerationUpperLimit(std::size_t /*_index*/,
                                              double /*_acceleration*/)
 {
@@ -395,6 +457,18 @@ void ZeroDofJoint::setAccelerationUpperLimit(std::size_t /*_index*/,
 double ZeroDofJoint::getAccelerationUpperLimit(std::size_t /*_index*/) const
 {
   return 0.0;
+}
+
+//==============================================================================
+void ZeroDofJoint::setAccelerationUpperLimits(const Eigen::VectorXd& /*upperLimits*/)
+{
+  // Do nothing
+}
+
+//==============================================================================
+Eigen::VectorXd ZeroDofJoint::getAccelerationUpperLimits() const
+{
+  return Eigen::VectorXd::Zero(0);
 }
 
 //==============================================================================
@@ -440,6 +514,18 @@ double ZeroDofJoint::getForceLowerLimit(std::size_t /*_index*/) const
 }
 
 //==============================================================================
+void ZeroDofJoint::setForceLowerLimits(const Eigen::VectorXd& /*lowerLimits*/)
+{
+  // Do nothing
+}
+
+//==============================================================================
+Eigen::VectorXd ZeroDofJoint::getForceLowerLimits() const
+{
+  return Eigen::VectorXd::Zero(0);
+}
+
+//==============================================================================
 void ZeroDofJoint::setForceUpperLimit(std::size_t /*_index*/, double /*_force*/)
 {
   // Do nothing
@@ -449,6 +535,18 @@ void ZeroDofJoint::setForceUpperLimit(std::size_t /*_index*/, double /*_force*/)
 double ZeroDofJoint::getForceUpperLimit(std::size_t /*_index*/) const
 {
   return 0.0;
+}
+
+//==============================================================================
+void ZeroDofJoint::setForceUpperLimits(const Eigen::VectorXd& /*upperLimits*/)
+{
+  // Do nothing
+}
+
+//==============================================================================
+Eigen::VectorXd ZeroDofJoint::getForceUpperLimits() const
+{
+  return Eigen::VectorXd::Zero(0);
 }
 
 //==============================================================================