+++ unittests/TestHelpers.hpp
@@ -508,6 +508,12 @@ struct PresentResourceRetriever : public dart::common::ResourceRetriever
     return true;
   }
 
+  std::string getFilePath(const dart::common::Uri& _uri) override
+  {
+    mGetFilePath.push_back(_uri.toString());
+    return _uri.toString();
+  }
+
   dart::common::ResourcePtr retrieve(const dart::common::Uri& _uri) override
   {
     mRetrieve.push_back(_uri.toString());
@@ -515,6 +521,7 @@ struct PresentResourceRetriever : public dart::common::ResourceRetriever
   }
 
   std::vector<std::string> mExists;
+  std::vector<std::string> mGetFilePath;
   std::vector<std::string> mRetrieve;
 };
 
@@ -527,6 +534,12 @@ struct AbsentResourceRetriever : public dart::common::ResourceRetriever
     return false;
   }
 
+  std::string getFilePath(const dart::common::Uri& _uri) override
+  {
+    mGetFilePath.push_back(_uri.toString());
+    return "";
+  }
+
   dart::common::ResourcePtr retrieve(const dart::common::Uri& _uri) override
   {
     mRetrieve.push_back(_uri.toString());
@@ -534,6 +547,7 @@ struct AbsentResourceRetriever : public dart::common::ResourceRetriever
   }
 
   std::vector<std::string> mExists;
+  std::vector<std::string> mGetFilePath;
   std::vector<std::string> mRetrieve;
 };